Dia 15 - Setup da Autenticação no Backend

+++ projeto/backend/src/xp/xp.controller.ts
@@ -1,14 +1,19 @@
-import { Controller, Get, Query, Param, Post, Body, Put, Delete } from '@nestjs/common';
+import { Controller, Get, Query, Param, Post, Body, Put, Delete, UseGuards, Request } from '@nestjs/common';
 import { XpService } from './xp.service';
 import { criarXpDto } from './dto/criarXp.dto';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { OptionalJwtAuthGuard } from 'src/auth/optional-jwt-auth.guard';
 
 @Controller('xp')
 export class XpController {
     constructor(private readonly xpService: XpService) {}
 
     // Deleta uma experiência pelo id
+    @UseGuards(JwtAuthGuard)
     @Delete('remover/:id')
-    removeXp(@Param('id') id: string){
+    async removeXp(@Param('id') id: string, @Request() req){
+        const xp = await this.xpService.getXp(Number(id), req.user.id);
+        if (!xp || xp.id_user !== req.user.id) return null;
         return this.xpService.removeXp(Number(id));
     }
 
@@ -20,28 +25,37 @@ export class XpController {
     }
 
     // Coleta as experiencias de um usuário com filtros (se tiver)
-    @Get('usuario/:id')
-    listarXpUser(@Param('id') id: string, @Query('busca') busca?: string, @Query('tag') tag?: string) {
+    @UseGuards(JwtAuthGuard)
+    @Get('usuario')
+    listarXpUser(@Request() req, @Query('busca') busca?: string, @Query('tag') tag?: string) {
         const tags = tag?.split(',');
-        return this.xpService.listarXpUser(Number(id), busca, tags);
+        return this.xpService.listarXpUser(req.user.id, busca, tags);
     }
 
     // Cria uma experiência
+    @UseGuards(JwtAuthGuard)
     @Post('criar')
-    criarXp(@Body() dto: criarXpDto){
+    criarXp(@Body() dto: criarXpDto, @Request() req){
+        dto.id_user = req.user.id;
         return this.xpService.criarXp(dto);
     }
 
     // Atualiza uma experiência
+    @UseGuards(JwtAuthGuard)
     @Put('atualizar')
-    atualizaXp(@Body() dto: {id: number, texto: string, contexto: string, tags: string[], pub: boolean, likes: number}){
+    async atualizaXp(@Body() dto: {id: number, texto: string, contexto: string, tags: string[], pub: boolean, likes: number}, @Request() req){
+        const xp = await this.xpService.getXp(dto.id, req.user.id);
+        if (!xp || xp.id_user !== req.user.id) return null;
         return this.xpService.atualizaXp(dto);
     }
 
     // Retorna uma experiência especifica pelo id
+    @UseGuards(OptionalJwtAuthGuard)
     @Get(':id')
-    getXp(@Param('id') id: string){
-        return this.xpService.getXp(Number(id));
+    getXp(@Param('id') id: string, @Request() req) {
+        const userId = req.user?.id; // estará preenchido se token for válido
+        return this.xpService.getXp(+id, userId);
     }
+
 
 }
